% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_CSS.R
\name{create_empty_logbins}
\alias{create_empty_logbins}
\title{Create log bin table}
\usage{
create_empty_logbins(max, logbase = exp(1))
}
\arguments{
\item{max}{numeric value of maximum size}

\item{logbase}{numeric value of log base to be used}
}
\value{
A tibble of log bins
}
\description{
Create an empty log bin table
}
