% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_SSD_params.R
\name{estimate_SSD_params}
\alias{estimate_SSD_params}
\title{Estimate species size distribuion (SSD)}
\usage{
estimate_SSD_params(
  x,
  meanlog_slope = 0.781,
  meanlog_intercept = -1.01,
  sdlog = 0.995
)
}
\arguments{
\item{x}{A tibble or dataframe with Lmax, a and b}
}
\value{
A tibble of meanlog and sdlog parameters of SSD
}
\description{
This function estimates a Species size distribuion based on
parameters from asymptotic mass
}
