% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_species.R
\docType{data}
\name{example_species}
\alias{example_species}
\title{Load species list}
\format{
A tibble of length 20 and one column
\describe{
  \item{species}{Species name (binomial nomenclature)}
}
}
\source{
<https://www.github.com/FreddieJH/sbss>
}
\usage{
example_species
}
\description{
This function loads an example species list dataset
}
\keyword{datasets}
