% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_lmax.R
\name{find_lmax}
\alias{find_lmax}
\title{Clean length-weight ests}
\usage{
find_lmax(df, spp_colname = "species_name", full_load = F)
}
\arguments{
\item{df}{dataframe or tibble, must contain a column of species names}

\item{spp_colname}{character, the name of the column containing the species names}

\item{full_load}{logical, if TRUE will pull values from Fishbase and Sealifebase}
}
\value{
A tibble of assymptotic length variables
}
\description{
This function cleans length-weight estimates from sealifebase and fishbase.
If no appropriate length-weight estimate is available for a species, the mean
for a lower taxonomic level will be used.
}
