% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_lw.R
\name{find_lw}
\alias{find_lw}
\title{Clean length-weight ests}
\usage{
find_lw(df, spp_colname = "species_name", full_load = T)
}
\arguments{
\item{df}{tibble or dataframe, must contain a column of species names}

\item{spp_colname}{character, the name of the column containing the species names}

\item{full_load}{logical, if TRUE will load from Fishbase and Sealifebase}
}
\value{
tibble of length-weight variables
}
\description{
This function cleans length-weight estimates from sealifebase and fishbase.
If no appropriate length-weight estimate is available for a species, the mean
for a lower taxonomic level will be used
}
