% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_CSS.R
\name{integrate_csd}
\alias{integrate_csd}
\title{Integrate community size distribution}
\usage{
integrate_csd(df, min, max, abundance_colname = "abundance")
}
\arguments{
\item{df}{table with cols of sdlog, meanlog and relative abundance}

\item{min}{integer, min body size limit of integration}

\item{max}{integer, max body size limit of integration}

\item{abundance_colname}{character string, name of the column containing relative abundance}
}
\value{
integral of function
}
\description{
Integrates over the summed species
}
