% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params.R
\name{load_lw}
\alias{load_lw}
\title{Obtain length-weight parameters from Fishbase and Sealifebase}
\usage{
load_lw()
}
\value{
A tibble of length-weight parameters
}
\description{
This function obtains length-weight parameters from fishbase and
sealifebase databases, utilising the rfishbase package.
}
