% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params.R
\name{load_taxalist}
\alias{load_taxalist}
\title{Load Sealifebase and Fishbase Taxa}
\usage{
load_taxalist()
}
\value{
A tibble of taxa
}
\description{
This function loads a tibble of Sealifebase and Fishbase taxa
}
