% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_lmax.R
\name{lookup_lmax}
\alias{lookup_lmax}
\title{Clean length-weight ests}
\usage{
lookup_lmax(spp_table, full_load = F)
}
\arguments{
\item{spp_table}{A tibble or dataframe with a column of species names ("species")}

\item{full_load}{logical, if TRUE will pull values from Fishbase and Sealifebase.}
}
\value{
A tibble of length-weight variables
}
\description{
This function cleans length-weight estimates from sealifebase and fishbase.
If no appropriate length-weight estimate is available for a species, the mean
for a lower taxonomic level will be used
}
