% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SSDs.R
\name{plot_ssd}
\alias{plot_ssd}
\title{Plot species size distributions}
\usage{
plot_ssd(data, log = F, logbase = NA, max_x = 100, max_y = 0.5, legend = T)
}
\arguments{
\item{data}{dataframe or tibble, containing the columns `meanlog`, and `sdlog`.}

\item{log}{logical, if true axes will be logged}

\item{logbase}{double, log base for the plotting}

\item{max_x}{integer, determining the limit of the x-axis}

\item{max_y}{integer, value determining the limit of the y-axis}

\item{legend}{logical, if TRUE legend of species names will be included}
}
\value{
A ggplot figure
}
\description{
This function plots species size distributions
}
