% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SSDs.R
\name{plot_SSDs}
\alias{plot_SSDs}
\title{Plot species size distributions}
\usage{
plot_SSDs(data, log = F, logbase = NA, max_x = 100, max_y = 0.5, legend = T)
}
\arguments{
\item{data}{Dataframe or tibble}

\item{log}{Logical argument for logged axes}

\item{logbase}{The chosen log base for the plotting}

\item{max_x}{Numeric value determining the limit of the x-axis}

\item{max_y}{Numeric value determining the limit of the y-axis}
}
\value{
A ggplot figure
}
\description{
This function plots species size distributions
}
