% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_CSS.R
\name{sum_ssd}
\alias{sum_ssd}
\title{Sum up species size distributions}
\usage{
sum_ssd(x, meanlog, sdlog, weight)
}
\arguments{
\item{x}{vector of quantiles}

\item{meanlog}{vector of lognormal meanlog parameters (mean)}

\item{sdlog}{vector of lognormal sdlog parameters (standard devaiation)}

\item{weight}{vector of weights (likely the relative abundance of the species)}
}
\value{
vector of probabilities
}
\description{
Weighted sum of log-normal species body size distributions
}
