% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_CSS.R
\name{calc_CSS}
\alias{calc_CSS}
\title{Calculate the commuinity size spectrum}
\usage{
calc_CSS(x, group_var)
}
\arguments{
\item{x}{A tibble or dataframe with Species, rel_abun, meanlog, sdlog, Mmax, and a grouping variable (e.g. site)}
}
\value{
A tibble of normalised and non-normalised abundance with body size (m)
}
\description{
This function estimates a size spectrum based on
}
