% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_CSS.R
\name{calc_css}
\alias{calc_css}
\title{Calculate the community size spectrum}
\usage{
calc_css(df, group_var, abundance_colname = "abundance")
}
\arguments{
\item{df}{A tibble or dataframe with Species, rel_abun, meanlog, sdlog, Mmax, and a grouping variable (e.g. site)}

\item{group_var}{A string of the desired grouping variable}

\item{abundance_colname}{character, name of the column containing abundance information}
}
\value{
A tibble of normalised and non-normalised abundance with body size (m)
}
\description{
This function estimates a size spectrum based on
}
