% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_CSS.R
\name{integrate_CSD}
\alias{integrate_CSD}
\title{Integrate community size distribution}
\usage{
integrate_CSD(data, min, max, abundance_colname = "abundance")
}
\arguments{
\item{min}{min limit of integration}

\item{max}{max limit of integration}

\item{sdlog}{table with}

\item{meanlog}{table with}

\item{rel_abun}{table with}
}
\value{
integral of function
}
\description{
Integrates over the summed species
}
