% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_lw.R
\name{lookup_lw}
\alias{lookup_lw}
\title{Clean length-weight ests}
\usage{
lookup_lw(spp_table)
}
\arguments{
\item{spp_lmax_tab}{A tibble or dataframe with a column of species names ("species_name") and a column of lmax ("species_lmax")}
}
\value{
A tibble of length-weight variables
}
\description{
This function cleans length-weight estimates from sealifebase and fishbase.
If no appropriate length-weight estimate is available for a species, the mean
for a lower taxonomic level will be used
}
