% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_CSS.R
\name{plot_CSS}
\alias{plot_CSS}
\title{Plot community size spectrum}
\usage{
plot_CSS(.df, normalised = T, logbase = 2)
}
\arguments{
\item{normalised}{Logical option of normalised (TRUE) or non-normalised (FALSE)}

\item{logbase}{The chosen logbase for the plotting}

\item{x}{Table of mass, normalised density, and density}
}
\value{
A ggplot figure
}
\description{
This function plots an abundance size spectrum, with the option for
normalised or non-normalised axes
}
