% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SSDs.R
\name{plot_SSDs}
\alias{plot_SSDs}
\title{Plot species size distributions}
\usage{
plot_SSDs(.df, log = F, logbase = NA, max_x = 100, max_y = 0.5)
}
\arguments{
\item{.df}{Dataframe or tibble}

\item{logbase}{The chosen logbase for the plotting}

\item{normalised}{Logical option of normalised (TRUE) or non-normalised (FALSE)}
}
\value{
A ggplot figure
}
\description{
This function plots species size distributions
}
